+++ server/src/main/kotlin/com/colecto/colecto/entity/Article.kt
@@ -1,6 +1,6 @@
 package com.colecto.colecto.entity
 
-data class Product (
+data class Article (
   val id: Long,
   val title: String,
   val icon: String,
