--- server/src/test/kotlin/com/colecto/colecto/ColectoApplicationTests.kt
@@ -1,15 +1,0 @@
-package com.colecto.colecto
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest
-
-@DataJpaTest
-@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
-class ColectoApplicationTests {
-
-    @Test
-    fun contextLoads() {
-        println("test")
-    }
-}
