+++ server/src/test/kotlin/com/colecto/colecto/IntegrationTest.kt
@@ -1,0 +1,8 @@
+package com.colecto.colecto
+
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@AutoConfigureMockMvc
+open class IntegrationTest
